refactor(timer): 调整正向计时器参数和调整书架及相关屏幕的onTimerClick回调签名 以及  整理了删除书籍确认对话框代码
- 使用selectedBookId替换newlyAddedBookId参数
- 统一将ForwardTimerScreen和SaveRecordScreen中传递的书籍ID参数由newlyAddedBookId改为selectedBookId
- 修改相关ViewModel和导航逻辑，确保selectedBookId正确传递和处理
- 在ForwardTimerViewModel中增加了详细日志，强化选书和异常处理的可观测性
- 优化了计时器停止逻辑，增加选中书籍的存在性检查和日志输出
- 调整书架及相关屏幕的onTimerClick回调签名，支持传递书籍ID参数以增强交互逻辑
- 整理了删除书籍确认对话框代码，提取成公共组件并在书架各详情页中集成
- 修改书架相关页面和组件，增加书籍删除确认对话框的显示和功能实现
- 统一完善UI层代码结构和导入，提升代码清晰度与可维护性

+++ app/src/main/java/com/anou/pagegather/ui/feature/bookshelf/BookShelfScreen.kt
@@ -12,7 +12,7 @@ fun BookShelfScreen(
     onNavigateToBookGroups: () -> Unit,
     onToBookViewClick: (Long) -> Unit,
     onToBookAddClick: () -> Unit,
-    onNavigateToTimer: () -> Unit = {},
+    onNavigateToTimer: (Long) -> Unit = {},
     onNavigateToQuickActions: () -> Unit = {},
     onNavigateToNoteEdit: ((Long, Long) -> Unit)? = null,  // 添加导航到笔记编辑页面的回调函数
     onNavigateToBookEdit: ((Long) -> Unit)? = null,  // 添加导航到书籍编辑页面的回调函数
